+++ packages/components/modules/messages/native/GroupDetailsPage/Members/index.tsx
@@ -1,7 +1,10 @@
 import { FC } from 'react'
 
-import { Button, FabButton } from '@baseapp-frontend/design-system/components/native/buttons'
-import { LoadingScreen } from '@baseapp-frontend/design-system/components/native/displays'
+import {
+  Button,
+  FabButton as DefaultFabButton,
+} from '@baseapp-frontend/design-system/components/native/buttons'
+import { LoadingScreen as DefaultLoadingScreen } from '@baseapp-frontend/design-system/components/native/displays'
 import { Text } from '@baseapp-frontend/design-system/components/native/typographies'
 import { View } from '@baseapp-frontend/design-system/components/native/views'
 import { useTheme } from '@baseapp-frontend/design-system/providers/native'
@@ -9,7 +12,7 @@ import { useTheme } from '@baseapp-frontend/design-system/providers/native'
 import { useRouter } from 'expo-router'
 
 import { CHAT_ROOM_PARTICIPANT_ROLES } from '../../../common'
-import MemberItem from './MemberItem'
+import DefaultMemberItem from './MemberItem'
 import { createStyles } from './styles'
 import { MembersProps } from './type'
 
@@ -19,6 +22,14 @@ const Members: FC<MembersProps> = ({
   isLoadingNext,
   hasNext,
   loadNext,
+  currentProfileIsAdmin = false,
+  groupId,
+  MemberItem = DefaultMemberItem,
+  MemberItemProps = {},
+  FabButton = DefaultFabButton,
+  FabButtonProps = {},
+  LoadingScreen = DefaultLoadingScreen,
+  LoadingScreenProps = {},
   roomId,
 }) => {
   const theme = useTheme()
@@ -43,6 +54,7 @@ const Members: FC<MembersProps> = ({
           iconSize={28}
           iconColor={theme.colors.primary.contrast}
           style={styles.addMemberButton}
+          {...FabButtonProps}
         />
         <Text
           variant="subtitle2"
@@ -61,13 +73,16 @@ const Members: FC<MembersProps> = ({
           return (
             <MemberItem
               key={edge.node.id ?? `member-edge-${index}`}
-              profileRef={edge.node.profile}
-              isAdmin={edge.node.role === CHAT_ROOM_PARTICIPANT_ROLES.admin}
+              groupMember={edge.node}
+              memberIsAdmin={edge.node.role === CHAT_ROOM_PARTICIPANT_ROLES.admin}
+              currentProfileIsAdmin={currentProfileIsAdmin}
+              groupId={groupId}
+              {...MemberItemProps}
             />
           )
         })}
         {isLoadingNext ? (
-          <LoadingScreen size="small" />
+          <LoadingScreen size="small" {...LoadingScreenProps} />
         ) : (
           hasNext && (
             <Button mode="text" size="medium" onPress={loadNext}>