+++ packages/components/modules/messages/native/GroupDetailsPage/Members/type.ts
@@ -1,4 +1,10 @@
+import { FC } from 'react'
+
+import { FabButtonProps } from '@baseapp-frontend/design-system/components/native/buttons'
+import { LoadingScreenProps } from '@baseapp-frontend/design-system/components/native/displays'
+
 import { MembersListFragment$data } from '../../../../../__generated__/MembersListFragment.graphql'
+import { MemberItemProps } from './MemberItem/type'
 
 export interface MembersProps {
   participantsCount?: number | null
@@ -6,5 +12,13 @@ export interface MembersProps {
   loadNext: () => void
   isLoadingNext: boolean
   hasNext?: boolean
+  currentProfileIsAdmin?: boolean
+  groupId: string
+  MemberItem?: FC<MemberItemProps>
+  MemberItemProps?: Partial<MemberItemProps>
+  FabButton?: FC<FabButtonProps>
+  FabButtonProps?: Partial<FabButtonProps>
+  LoadingScreen?: FC<LoadingScreenProps>
+  LoadingScreenProps?: Partial<LoadingScreenProps>
   roomId?: string
 }