+++ packages/components/modules/messages/native/GroupDetailsPage/index.tsx
@@ -18,18 +18,28 @@ import {
   useCheckIsAdmin,
 } from '../../common'
 import { useGroupChatCreate } from '../../common/context/GroupChatProvider'
-import { LeaveGroupDialog } from '../__shared__/LeaveGroupDialog'
-import GroupProfile from './GroupProfile'
-import Members from './Members'
-import Options from './Options'
+import { LeaveGroupDialog as DefaultLeaveGroupDialog } from '../__shared__/LeaveGroupDialog'
+import { useMessagesListSubscription } from '../graphql/subscriptions/useMessagesListSubscription'
+import DefaultGroupProfile from './GroupProfile'
+import DefaultMembers from './Members'
+import DefaultOptions from './Options'
 import { GroupDetailsPageProps } from './type'
 
-const GroupDetailsPage: FC<GroupDetailsPageProps> = ({ roomId }) => {
+const GroupDetailsPage: FC<GroupDetailsPageProps> = ({
+  roomId,
+  LeaveGroupDialog = DefaultLeaveGroupDialog,
+  LeaveGroupDialogProps = {},
+  GroupProfile = DefaultGroupProfile,
+  GroupProfileProps = {},
+  Members = DefaultMembers,
+  MembersProps = {},
+  Options = DefaultOptions,
+  OptionsProps = {},
+}) => {
   const router = useRouter()
   const [openConfirmLeaveGroupDialog, setOpenConfirmLeaveGroupDialog] = useState(false)
   const { currentProfile } = useCurrentProfile()
-  const [commitMarkAsRead, isMutationInFlight] = useArchiveChatRoomMutation()
-
+  const [commitArchiveRoom, isMutationInFlight] = useArchiveChatRoomMutation()
   const groups = useGroupChatCreate()
 
   const { chatRoom: group } = useLazyLoadQuery<GroupDetailsQueryType>(
@@ -53,7 +63,7 @@ const GroupDetailsPage: FC<GroupDetailsPageProps> = ({ roomId }) => {
     MembersListFragment$key
   >(MembersListFragment, group)
   const members = data?.participants
-  const { isSoleAdmin } = useCheckIsAdmin(members)
+  const { isSoleAdmin, isAdmin: currentProfileIsAdmin } = useCheckIsAdmin(members)
 
   const handleLoadMoreMembers = () => {
     if (hasNext && !isLoadingNext) {
@@ -63,7 +73,7 @@ const GroupDetailsPage: FC<GroupDetailsPageProps> = ({ roomId }) => {
 
   const handleArchiveChat = () => {
     if (currentProfile?.id && roomId) {
-      commitMarkAsRead({
+      commitArchiveRoom({
         variables: {
           input: {
             roomId,
@@ -74,6 +84,8 @@ const GroupDetailsPage: FC<GroupDetailsPageProps> = ({ roomId }) => {
       })
     }
   }
+
+  useMessagesListSubscription(roomId, currentProfile?.id ?? '')
 
   return (
     <View style={{ flexGrow: 1, flex: 1 }}>
@@ -94,22 +106,27 @@ const GroupDetailsPage: FC<GroupDetailsPageProps> = ({ roomId }) => {
           roomId={roomId}
           removingParticipantId={currentProfile?.id}
           isSoleAdmin={isSoleAdmin}
+          {...LeaveGroupDialogProps}
         />
       )}
       <ScrollView>
-        <GroupProfile group={group} />
+        <GroupProfile group={group} {...GroupProfileProps} />
         <Members
           participantsCount={group?.participantsCount}
           members={members}
           loadNext={handleLoadMoreMembers}
           isLoadingNext={isLoadingNext}
           hasNext={hasNext}
+          currentProfileIsAdmin={currentProfileIsAdmin}
+          groupId={roomId}
+          {...MembersProps}
         />
         <Options
           isArchiveMutationInFlight={isMutationInFlight}
           handleArchiveChat={handleArchiveChat}
           handleLeaveGroup={() => setOpenConfirmLeaveGroupDialog(true)}
           isArchived={!!group?.isArchived}
+          {...OptionsProps}
         />
       </ScrollView>
     </View>