+++ packages/components/modules/messages/native/RoomsList/ChatCard/ChatCardComponent/index.tsx
@@ -110,7 +110,7 @@ const ChatCardComponent: FC<ChatCardComponentProps> = ({ roomRef, isArchived }) 
       router.push(`/group-details/${roomId}`)
       return
     }
-    console.log('Not implemented yet.')
+    router.push(`/single-chat-details/${roomId}`)
   }
   const handleGoToProfile = () => {
     bottomDrawerRef.current?.close()