+++ packages/components/modules/navigations/web/__shared__/NavToggleButton/types.ts
@@ -1,3 +1,10 @@
+import { FC } from 'react'
+
+import { ChevronIconProps } from '@baseapp-frontend/design-system/components/web/icons'
+
 import { IconButtonProps } from '@mui/material/IconButton'
 
-export interface NavToggleButtonProps extends IconButtonProps {}
+export interface NavToggleButtonProps extends IconButtonProps {
+  ChevronIcon?: FC<ChevronIconProps>
+  ChevronIconProps?: Partial<ChevronIconProps>
+}