+++ packages/wagtail/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @baseapp-frontend/wagtail
 
+## 1.0.50
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/design-system@1.2.4
+
+## 1.0.49
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/design-system@1.2.3
+
+## 1.0.48
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/design-system@1.2.2
+
+## 1.0.47
+
+### Patch Changes
+
+- Updated dependencies
+  - @baseapp-frontend/design-system@1.2.1
+
 ## 1.0.46
 
 ### Patch Changes